LoadingPage moved to common
