--- src/features/LoadingPage/styled.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const Header = styled.header`
-    font-weight: 600;
-    font-size: 36px;
-    line-height: 1.2;
-    color: ${({ theme }) => theme.color.woodsmoke};
-`;
